Fix lighting system review issues
1. Fix lerpColor unsigned overflow: Cast uint8_t operands to float
   before subtraction to prevent wrap-around when b < a during
   dusk/night color transitions.

2. Fix const_cast config mutation: Add setLightingEnabled() method
   to LightingSystem, replacing the const_cast hack in Engine.cpp
   that bypassed setConfig() and its m_needsRecalc trigger.

3. Fix O(columns × chunks) skylight scan: Replace the inner loop
   that iterated all loaded chunks per column with direct
   setLight/getLight calls (O(1) per tile via chunk coordinate
   computation).

4. Fix syncChunksWithWorld quadratic removal: Iterate light map
   chunk keys directly via getChunkPositions() instead of scanning
   the full world coordinate range.

5. Fix flat lighting dead code: Remove unused overlayR/G/B colored
   light branch that computed per-channel values but never used
   them. Simplify to direct max-darkness alpha overlay.

6. Clean up variable shadowing: Rename loop variables in
   propagateSkylight to avoid shadowing outer loop names.

7. Remove dead tile light source placeholder: Delete the unused
   visible-area padding code block in collectLightSources().

8. Change loadedSet from map<bool> to unordered_set.

9. Add regression test for lerpColor with decreasing channels
   (dusk→night transition where all RGB channels decrease).

All 372 tests pass (54 lighting + 318 existing).

+++ src/lighting/DayNightCycle.hpp
@@ -139,9 +139,9 @@ private:
     static TileLight lerpColor(const TileLight& a, const TileLight& b, float t) {
         t = std::max(0.0f, std::min(1.0f, t));
         return TileLight(
-            static_cast<uint8_t>(a.r + (b.r - a.r) * t),
-            static_cast<uint8_t>(a.g + (b.g - a.g) * t),
-            static_cast<uint8_t>(a.b + (b.b - a.b) * t)
+            static_cast<uint8_t>(static_cast<float>(a.r) + (static_cast<float>(b.r) - static_cast<float>(a.r)) * t),
+            static_cast<uint8_t>(static_cast<float>(a.g) + (static_cast<float>(b.g) - static_cast<float>(a.g)) * t),
+            static_cast<uint8_t>(static_cast<float>(a.b) + (static_cast<float>(b.b) - static_cast<float>(a.b)) * t)
         );
     }
